+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Enhancements
   * Add **Dockerfile** support
   * Add `--rebuild` and `--pull` options for the `azk start` or `azk restart` commands. It forces rebuild or image pulling and reprovision before start an instance (default: false).
+  * Added support to hide the help command options
 
 ## v0.8.3