+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Add **Dockerfile** support
   * Add `--rebuild` and `--pull` options for the `azk start` or `azk restart` commands. It forces rebuild or image pulling and reprovision before start an instance (default: false).
   * Added support to hide the help command options
+  * Adding option `--quiet` to support `non-interactive` mode in commands.
 
 ## v0.8.3